Add resource blueprint from ember-cli-legacy-blueprints
The idea is to reuse as much as we can from the original blueprints of
Ember and, later, we're going to be overrding them.

The goal of `ember g resource` is that it would create the following
files:
- Three routes (create, list and edit)
- One model (resource)
- Two components (form-resource and list-resource)
- Three controllers (Would we need query params?)
- Test files

In a second phase we may suggest to create also the needed SCSS files.