Add controllers when creating a new resource
Motivation: I'm creating the path that we would need for every view.
There are some specific actions that we would have in some specific
routes:
  - saveModel
  - queryModel

There are two strategies of adition of these actions: inside the Router
or in the Controller. As I thought that the community is more able to
use "controller first" to bind actions available in the view and that
they are simpler to get, I've decided to use this approach.

So, in the following improvements, we'll start to add functionality to
the controller, creating some basic ones (the candidates are
ListController and FormController). As we haven't decided yet, Mixins
and Services to solve this DRY problem are also posible solutions.